Exports the angles of the selected object and its outline.
This script will export the outline of the selected object from 3ds max
and the outline of said object in the folowing dimmensions 10 ,20, 30,
40, 50 pixel square.

The main problem right now with this process is the time, I realized
that it needs pro optimizer to be applied to the object and the vert
count must be under 20 thousand. the optimal would be under 10 thousand,
this way it would shorten the total processing time to around 5 minutes.